requirements.txt
# Django
django
django-cors-headers
django-filter
daphne
channels
channels-redis
celery
django-celery-results
django-crontab

# Django REST
djangorestframework
djangorestframework-simplejwt
drf-flex-fields
drf-extra-fields
drf-writable-nested

# Swagger/OpenAPI
drf-spectacular

# Database
psycopg2

# Debug
debugpy
django-silk

# Environment variables
python-dotenv

# Image field
Pillow

# For calling external APIs
requests

# WSGI Server
gunicorn

# benchmark
locust
tqdm